ajout Multithreading functions
